+++ src/components/icon/index.tsx
@@ -29,6 +29,9 @@ const ICONS = {
     '<g fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="1.5"><rect width="3" height="3" x="3" y="4" rx=".75"/><path d="M10 5.5h11"/><rect width="3" height="3" x="3" y="10.5" rx=".75"/><path d="M10 12h11"/><rect width="3" height="3" x="3" y="17" rx=".75"/><path d="M10 18.5h11"/></g>',
   palette:
     '<path fill="currentColor" d="M12 22q-2.05 0-3.875-.788t-3.187-2.15t-2.15-3.187T2 12q0-2.075.813-3.9t2.2-3.175T8.25 2.788T12.2 2q2 0 3.775.688t3.113 1.9t2.125 2.875T22 11.05q0 2.875-1.75 4.413T16 17h-1.85q-.225 0-.312.125t-.088.275q0 .3.375.863t.375 1.287q0 1.25-.687 1.85T12 22m-5.5-9q.65 0 1.075-.425T8 11.5t-.425-1.075T6.5 10t-1.075.425T5 11.5t.425 1.075T6.5 13m3-4q.65 0 1.075-.425T11 7.5t-.425-1.075T9.5 6t-1.075.425T8 7.5t.425 1.075T9.5 9m5 0q.65 0 1.075-.425T16 7.5t-.425-1.075T14.5 6t-1.075.425T13 7.5t.425 1.075T14.5 9m3 4q.65 0 1.075-.425T19 11.5t-.425-1.075T17.5 10t-1.075.425T16 11.5t.425 1.075T17.5 13"/>',
+  imageSparkle:
+    '<g fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="1.5"><rect width="16" height="15" x="3" y="6" rx="2"/><circle cx="8.25" cy="11.25" r="1.25"/><path d="m4.5 18l3.75-3.75a1.5 1.5 0 0 1 2.121 0L14 17.879l1.129-1.129a1.5 1.5 0 0 1 2.121 0L19 18.5M19.5 2v4M21.5 4h-4"/></g>',
+  warning: '<path fill="currentColor" d="M1 21L12 2l11 19zm10-3h2v-2h-2zm0-4h2v-4h-2z"/>',
   speed:
     '<path fill="currentColor" d="M10.45 15.5q.625.625 1.575.588T13.4 15.4L19 7l-8.4 5.6q-.65.45-.712 1.362t.562 1.538M5.1 20q-.55 0-1.012-.238t-.738-.712q-.65-1.175-1-2.437T2 14q0-2.075.788-3.9t2.137-3.175T8.1 4.788T12 4q2.05 0 3.85.775T19 6.888t2.15 3.125t.825 3.837q.025 1.375-.312 2.688t-1.038 2.512q-.275.475-.737.713T18.874 20z"/>',
   checkCircle: